Initial Commit
This project is simply to create support for dialogue in an engine
to create more story driven experiences. As of right now the project
just contains the main basic game engine which I created today. This
project is also to get myself familiar with the PEP-8 standards for
coding, and that is why there are extensive comments. Every file scored
at least a 9/10 using pylint, with most getting 10.